backend/requirements.txt
azure-core==1.39.0
azure-cosmos==4.15.0
azure-storage-blob==12.28.0
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
Flask==3.1.3
flask-cors==6.0.2
gunicorn
idna==3.12
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==3.0
python-dotenv==1.2.2
requests==2.33.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
